University Lecture/Instructor Evaluations by Students at ETH

University lecture evaluations by students at ETH Zurich, anonymized for privacy protection. This is an interesting “medium” sized example of a partially nested mixed effect model.

Format

A data frame with 73421 observations on the following 7 variables.

s

a factor with levels 1:2972 denoting individual students.

d

a factor with 1128 levels from 1:2160, denoting individual professors or lecturers.

% ("d": \dQuote{Dozierende} in German)
studage

an ordered factor with levels 2 < 4 < 6 < 8, denoting student's “age” measured in the semester number the student has been enrolled.

lectage

an ordered factor with 6 levels, 1 < 2 < ... < 6, measuring how many semesters back the lecture rated had taken place.

service

a binary factor with levels 0 and 1; a lecture is a “service”, if held for a different department than the lecturer's main one.

dept

a factor with 14 levels from 1:15, using a random code for the department of the lecture.

y

a numeric vector of ratings of lectures by the students, using the discrete scale 1:5, with meanings of ‘poor’ to ‘very good’.

Each observation is one student's rating for a specific lecture (of one lecturer, during one semester in the past).

Details

The main goal of the survey is to find “the best liked prof”, according to the lectures given. Statistical analysis of such data has been the basis for a (student) jury selecting the final winners.

The present data set has been anonymized and slightly simplified on purpose.

Examples

str(InstEval)
#> 'data.frame':	73421 obs. of  7 variables:
#>  $ s      : Factor w/ 2972 levels "1","2","3","4",..: 1 1 1 1 2 2 3 3 3 3 ...
#>  $ d      : Factor w/ 1128 levels "1","6","7","8",..: 525 560 832 1068 62 406 3 6 19 75 ...
#>  $ studage: Ord.factor w/ 4 levels "2"<"4"<"6"<"8": 1 1 1 1 1 1 1 1 1 1 ...
#>  $ lectage: Ord.factor w/ 6 levels "1"<"2"<"3"<"4"<..: 2 1 2 2 1 1 1 1 1 1 ...
#>  $ service: Factor w/ 2 levels "0","1": 1 2 1 2 1 1 2 1 1 1 ...
#>  $ dept   : Factor w/ 14 levels "15","5","10",..: 14 5 14 12 2 2 13 3 3 3 ...
#>  $ y      : int  5 2 5 3 2 4 4 5 5 4 ...

head(InstEval, 16)
#>    s    d studage lectage service dept y
#> 1  1 1002       2       2       0    2 5
#> 2  1 1050       2       1       1    6 2
#> 3  1 1582       2       2       0    2 5
#> 4  1 2050       2       2       1    3 3
#> 5  2  115       2       1       0    5 2
#> 6  2  756       2       1       0    5 4
#> 7  3    7       2       1       1   11 4
#> 8  3   13       2       1       0   10 5
#> 9  3   36       2       1       0   10 5
#> 10 3  140       2       1       0   10 4
#> 11 3  409       2       2       0   10 4
#> 12 3  444       2       2       0   10 4
#> 13 3  494       2       1       1    9 4
#> 14 3  625       2       2       0   10 3
#> 15 3  696       2       2       1    9 2
#> 16 3 1056       2       2       1    8 4
xtabs(~ service + dept, InstEval)
#>        dept
#> service   15    5   10   12    6    7    4    8    9   14    1    3   11    2
#>       0 2466 3576 4343 6209 3772 1601 4518  578 4224 2606 1260 3550 1711 1224
#>       1  826  214  365 3319 4325  919 2207 3848 2400 1328 1372 1199 6863 2598