NelderMead-class {lme4}	R Documentation

Class "NelderMead" of Nelder-Mead optimizers and its Generator

Description

Class "NelderMead" is a reference class for a Nelder-Mead simplex optimizer allowing box constraints on the parameters and using reverse communication.

The NelderMead() function conveniently generates such objects.

Usage

NelderMead(...)

Arguments

...
Argument list (see Note below).

Methods

NelderMead$new(lower, upper, xst, x0, xt)

Create a new NelderMead object

Extends

All reference classes extend and inherit methods from "envRefClass".

Note

This is the default optimizer for the second stage of glmer and nlmer fits. We found that it was more reliable and often faster than more sophisticated optimizers.

Arguments to NelderMead() and the new method must be named arguments:

lower

numeric vector of lower bounds - elements may be -Inf.

upper

numeric vector of upper bounds - elements may be Inf.

xst

numeric vector of initial step sizes to establish the simplex - all elements must be non-zero.

x0

numeric vector of starting values for the parameters.

xt

numeric vector of tolerances on the parameters.

References

Based on code in the NLopt collection.

See Also

Nelder_Mead, the typical “constructor”. Further, glmer, nlmer

Examples

showClass("NelderMead")

[Package lme4 version 1.1-35.1 Index]