factorize {lme4}	R Documentation

Attempt to convert grouping variables to factors

Description

If variables within a data frame are not factors, try to convert them. Not intended for end-user use; this is a utility function that needs to be exported, for technical reasons.

Usage

factorize(x,frloc,char.only=FALSE)

Arguments

x
a formula

frloc
a data frame

char.only
(logical) convert only character variables to factors?

Value

a copy of the data frame with factors converted

[Package lme4 version 1.1-35.1 Index]