fortify {lme4}	R Documentation

add information to data based on a fitted model

Description

add information to data based on a fitted model

Usage

  fortify.merMod(model, data = getData(model),
    ...)

Arguments

model
fitted model

data
original data set, if needed

...
additional arguments

Details

fortify is a function defined in the ggplot2 package, q.v. for more details. fortify is not defined here, and fortify.merMod is defined as a function rather than an S3 method, to avoid (1) inducing a dependency on ggplot2 or (2) masking methods from ggplot2. This is currently an experimental feature.

[Package lme4 version 1.1-19 Index]