Class "glmFamily" - a reference class for family

This class is a wrapper class for family objects specifying a distibution family and link function for a generalized linear model (glm). The reference class contains an external pointer to a C++ object representing the class. For common families and link functions the functions in the family are implemented in compiled code so they can be accessed from other compiled code and for a speed boost.

Note

Objects from this reference class correspond to objects in a C++ class. Methods are invoked on the C++ class using the external pointer in the Ptr field. When saving such an object the external pointer is converted to a null pointer, which is why there is a redundant field ptr that is an active-binding function returning the external pointer. If the Ptr field is a null pointer, the external pointer is regenerated for the stored family field.

Extends

All reference classes extend and inherit methods from "envRefClass".

See also

Examples

str(glmFamily$new(family=poisson()))
#> Reference class 'glmFamily' [package "lme4"] with 2 fields
#>  $ Ptr   :<externalptr> 
#>  $ family:List of 13
#>   ..$ family    : chr "poisson"
#>   ..$ link      : chr "log"
#>   ..$ linkfun   :function (mu)  
#>   ..$ linkinv   :function (eta)  
#>   ..$ variance  :function (mu)  
#>   ..$ dev.resids:function (y, mu, wt)  
#>   ..$ aic       :function (y, n, mu, wt, dev)  
#>   ..$ mu.eta    :function (eta)  
#>   ..$ initialize:  expression({  if (any(y < 0))  stop("negative values not allowed for the 'Poisson' family")  n <- rep.int(1, nobs| __truncated__
#>   ..$ validmu   :function (mu)  
#>   ..$ valideta  :function (eta)  
#>   ..$ simulate  :function (object, nsim)  
#>   ..$ dispersion: num 1
#>   ..- attr(*, "class")= chr "family"
#>  and 23 methods, of which 9 are  possibly relevant:
#>    aic, devResid, link, linkInv, muEta, ptr, setTheta, theta, variance