merPredD {lme4}	R Documentation

Generator object for the merPredD class

Description

The generator object for the merPredD reference class. Such an object is primarily used through its new method.

Usage

merPredD(...)

Arguments

...
List of arguments (see Note).

Note

merPredD(...) is a short form of new("merPredD", ...) to create a new merPredD object and the ... must be named arguments, (X, Zt, Lambdat, Lind, theta,n):

X:

dense model matrix for the fixed-effects parameters, to be stored in the X field.

Zt:

transpose of the sparse model matrix for the random effects. It is stored in the Zt field.

Lambdat:

transpose of the sparse lower triangular relative variance factor (stored in the Lambdat field).

Lind:

integer vector of the same length as the x slot in the Lambdat field. Its elements should be in the range 1 to the length of the theta field.

theta:

numeric vector of variance component parameters (stored in the theta field).

n:

sample size, usually nrow(X).

See Also

The class definition, merPredD, also for examples.

[Package lme4 version 1.1-35.1 Index]