Class "merPredD" - a Dense Predictor Reference Class

A reference class (see mother class definition "envRefClass" for a mixed-effects model predictor module with a dense model matrix for the fixed-effects parameters. The reference class is associated with a C++ class of the same name. As is customary, the generator object, merPredD, for the class has the same name as the class.

Note

Objects from this reference class correspond to objects in a C++ class. Methods are invoked on the C++ class object using the external pointer in the Ptr field. When saving such an object the external pointer is converted to a null pointer, which is why there are redundant fields containing enough information as R objects to be able to regenerate the C++ object. The convention is that a field whose name begins with an upper-case letter is an R object and the corresponding field, whose name begins with the lower-case letter is a method. References to the external pointer should be through the method, not directly through the Ptr field.

See also

Examples

showClass("merPredD")
#> Class "merPredD" [package "lme4"]
#> 
#> Slots:
#>                   
#> Name:       .xData
#> Class: environment
#> 
#> Extends: 
#> Class "envRefClass", directly
#> Class ".environment", by class "envRefClass", distance 2
#> Class "refClass", by class "envRefClass", distance 2
#> Class "environment", by class "envRefClass", distance 3, with explicit coerce
#> Class "refObject", by class "envRefClass", distance 3
pp <- slot(lmer(Yield ~ 1|Batch, Dyestuff), "pp")
stopifnot(is(pp, "merPredD"))
str(pp) # an overview of all fields and methods' names.
#> Reference class 'merPredD' [package "lme4"] with 18 fields
#>  $ Lambdat:Formal class 'dgCMatrix' [package "Matrix"] with 6 slots
#>   .. ..@ i       : int [1:6] 0 1 2 3 4 5
#>   .. ..@ p       : int [1:7] 0 1 2 3 4 5 6
#>   .. ..@ Dim     : int [1:2] 6 6
#>   .. ..@ Dimnames:List of 2
#>   .. .. ..$ : NULL
#>   .. .. ..$ : NULL
#>   .. ..@ x       : num [1:6] 0.848 0.848 0.848 0.848 0.848 ...
#>   .. ..@ factors : list()
#>  $ LamtUt :Formal class 'dgCMatrix' [package "Matrix"] with 6 slots
#>   .. ..@ i       : int [1:30] 0 0 0 0 0 1 1 1 1 1 ...
#>   .. ..@ p       : int [1:31] 0 1 2 3 4 5 6 7 8 9 ...
#>   .. ..@ Dim     : int [1:2] 6 30
#>   .. ..@ Dimnames:List of 2
#>   .. .. ..$ : NULL
#>   .. .. ..$ : chr [1:30] "1" "2" "3" "4" ...
#>   .. ..@ x       : num [1:30] 0.848 0.848 0.848 0.848 0.848 ...
#>   .. ..@ factors : list()
#>  $ Lind   : int [1:6] 1 1 1 1 1 1
#>  $ Ptr    :<externalptr> 
#>  $ RZX    : num [1:6, 1] 1.98 1.98 1.98 1.98 1.98 ...
#>  $ Ut     :Formal class 'dgCMatrix' [package "Matrix"] with 6 slots
#>   .. ..@ i       : int [1:30] 0 0 0 0 0 1 1 1 1 1 ...
#>   .. ..@ p       : int [1:31] 0 1 2 3 4 5 6 7 8 9 ...
#>   .. ..@ Dim     : int [1:2] 6 30
#>   .. ..@ Dimnames:List of 2
#>   .. .. ..$ : chr [1:6] "A" "B" "C" "D" ...
#>   .. .. ..$ : chr [1:30] "1" "2" "3" "4" ...
#>   .. ..@ x       : num [1:30] 1 1 1 1 1 1 1 1 1 1 ...
#>   .. ..@ factors : list()
#>  $ Utr    : num [1:6] 6384 6481 6634 6354 6787 ...
#>  $ V      : num [1:30, 1] 1 1 1 1 1 1 1 1 1 1 ...
#>  $ VtV    : num [1, 1] 30
#>  $ Vtr    : num 45825
#>  $ X      : num [1:30, 1] 1 1 1 1 1 1 1 1 1 1 ...
#>   ..- attr(*, "dimnames")=List of 2
#>   .. ..$ : chr [1:30] "1" "2" "3" "4" ...
#>   .. ..$ : chr "(Intercept)"
#>   ..- attr(*, "assign")= int 0
#>   ..- attr(*, "msgScaleX")= chr(0) 
#>  $ Xwts   : num [1:30] 1 1 1 1 1 1 1 1 1 1 ...
#>  $ Zt     :Formal class 'dgCMatrix' [package "Matrix"] with 6 slots
#>   .. ..@ i       : int [1:30] 0 0 0 0 0 1 1 1 1 1 ...
#>   .. ..@ p       : int [1:31] 0 1 2 3 4 5 6 7 8 9 ...
#>   .. ..@ Dim     : int [1:2] 6 30
#>   .. ..@ Dimnames:List of 2
#>   .. .. ..$ : chr [1:6] "A" "B" "C" "D" ...
#>   .. .. ..$ : chr [1:30] "1" "2" "3" "4" ...
#>   .. ..@ x       : num [1:30] 1 1 1 1 1 1 1 1 1 1 ...
#>   .. ..@ factors : list()
#>  $ beta0  : num 0
#>  $ delb   : num 1528
#>  $ delu   : num [1:6] -20.755 0.461 33.669 -27.212 66.877 ...
#>  $ theta  : num 0.848
#>  $ u0     : num [1:6] 0 0 0 0 0 0
#>  and 45 methods, of which 31 are  possibly relevant:
#>    CcNumer, L, P, RX, RXdiag, RXi, b, beta, copy#envRefClass, initialize,
#>    initializePtr, installPars, ldL2, ldRX2, linPred, ptr, setBeta0, setDelb,
#>    setDelu, setTheta, setZt, solve, solveU, sqrL, u, unsc, updateDecomp,
#>    updateL, updateLamtUt, updateRes, updateXwts