mkNewReTrms {lme4}	R Documentation

Make new random effect terms for prediction

Description

Rebuilds the random effect terms

Usage

  mkNewReTrms(object, newdata, re.form=NULL,
                     na.action=na.pass,
                     allow.new.levels=FALSE,
                     sparse = max(lengths(orig.random.levs)) > 100)

Arguments

object
a fitted merMod object

newdata
data frame for which to create new RE terms object

re.form
(formula, NULL, or NA) specify which random effects to condition on when predicting. If NULL, include all random effects; if NA or ~0, include no random effects

na.action
function determining what should be done with missing values for fixed effects in newdata

allow.new.levels
(logical) if new levels (or NA values) in newdata are allowed. If FALSE (default), such new values in newdata will trigger an error; if TRUE, then the prediction will use the unconditional (population-level) values for data with previously unobserved levels (or NAs)

sparse
generate sparse contrast matrices?

Value

a list with components

Zt
transpose of the sparse model matrix for the random effects

theta
initial values of the covariance parameters

Lind
an integer vector of indices determining the mapping of the elements of the theta vector to the "x" slot of Lambdat

Gp
a vector indexing the association of elements of the conditional mode vector with random-effect terms; if nb is the vector of numbers of conditional modes per term (i.e. number of groups times number of effects per group), Gp is c(0,cumsum(nb)) (and conversely nb is diff(Gp))

lower
lower bounds on the covariance parameters

Lambdat
transpose of the sparse relative covariance factor

flist
list of grouping factors used in the random-effects terms

cnms
a list of column names of the random effects according to the grouping factors

Ztlist
list of components of the transpose of the random-effects model matrix, separated by random-effects term

nl
names of the terms (in the same order as Zt, i.e. reflecting the reorder.terms argument)

Note

mkNewReTrms is used in the context of prediction, to generate a new "random effects terms" object from an already fitted model

See Also

Other utilities: findbars, mkRespMod, nlformula, nobars, subbars. getME can retrieve these components from a fitted model, although their values and/or forms may be slightly different in the final fitted model from their original values as returned from mkReTrms.

Examples

fm1 <- lmer(Reaction ~ Days + (Days | Subject), sleepstudy)
nd <- data.frame(Days = 5, Subject = "new")
try(mkNewReTrms(fm1, newdata = nd))
t1 <- mkNewReTrms(fm1, newdata = nd, allow.new.levels = TRUE)

[Package lme4 version 2.0-1 Index]