mkRespMod {lme4}	R Documentation

Create an lmerResp, glmResp or nlsResp instance

Description

Create an lmerResp, glmResp or nlsResp instance

Usage

  mkRespMod(fr, REML = NULL, family = NULL, nlenv = NULL,
    nlmod = NULL, ...)

Arguments

fr
a model frame

REML
logical scalar, value of REML for an lmerResp instance

family
the optional glm family (glmResp only)

nlenv
the nonlinear model evaluation environment (nlsResp only)

nlmod
the nonlinear model function (nlsResp only)

...
where to look for response information if fr is missing. Can contain a model response, y, offset, offset, and weights, weights.

Value

an lmerResp or glmResp or nlsResp instance

See Also

Other utilities: findbars, mkReTrms, nlformula, nobars, subbars

[Package lme4 version 1.1-31 Index]