mkSimulateTemplate {lme4}	R Documentation

Make templates suitable for guiding mixed model simulations

Description

Make data and parameter templates suitable for guiding mixed model simulations, by specifying a model formula and other information (EXPERIMENTAL). Most useful for simulating balanced designs and for getting started on unbalanced simulations.

Usage

mkParsTemplate(formula, data)
mkDataTemplate(formula, data, nGrps = 2, nPerGrp = 1, rfunc = NULL, ...)

Arguments

formula
A mixed model formula (see lmer).

data
A data frame containing the names in formula.

nGrps
Number of levels of a grouping factor.

nPerGrp
Number of observations per level.

rfunc
Function for generating covariate data (e.g. rnorm.

...
Additional parameters for rfunc.

See Also

These functions are designed to be used with simulate.merMod.

[Package lme4 version 1.1-35.1 Index]