namedList {lme4}	R Documentation

Self-naming list function

Description

this function takes a list of arguments and combines them into a list; any unnamed arguments are automatically named to match their symbols. The tibble::lst() function offers similarly functionality.

Usage

namedList(...)

Arguments

...
comma-separated arguments

Examples

a <- 1
b <- 2
c <- 3
str(namedList(a, b, d = c))

[Package lme4 version 1.1-35.1 Index]