rePos-class {lme4}	R Documentation

Class "rePos"

Description

A reference class for determining the positions in the random-effects vector that correspond to particular random-effects terms in the model formula

Extends

All reference classes extend and inherit methods from "envRefClass".

Examples

showClass("rePos")

[Package lme4 version 1.1-35.1 Index]