reformulas_imports {lme4}	R Documentation

(deprecated) formula-processing functions

Description

Formula-processing functions that have moved to reformulas

Usage

subbars(...)

nobars(...)

findbars(...)

mkReTrms(...)

isNested(...)

expandDoubleVerts(...)

reOnly(...)

Arguments

...
arguments to be passed through to functions in the reformulas package

Details

All of these functions are now located in the reformulas package. The versions here are 'shims' that throw a warning and call the reformulas version. (reOnly will not be fully fixed in reformulas until version 0.4.5 is on CRAN.)

[Package lme4 version 2.0-1 Index]