sigma {lme4}	R Documentation

Extract Residual Standard Deviation 'Sigma'

Description

Extract the estimated standard deviation of the errors, the “residual standard deviation” (also misnamed the “residual standard error”), from a fitted model of class merMod.

Usage

## S3 method for class 'merMod'
sigma(object, ...)

Arguments

object
a fitted model.

...
additional, optional arguments, passed from or to methods. (None currently in our two methods.)

Details

In general the dispersion parameter (which we call sigma) is the square root of the constant multiplier in the 'variance function' provided by the family functions.

For Gaussian fits, gaussian()$variance = rep.int(1, length(mu)) and it's known that the variance is σ^{2}. Hence, the constant multiplier of the variance function in this case is σ^{2}, so we say sigma is √{σ}.

For Gamma fits, Gamma()$variance = mu^2 where mu = scale*shape. The known variance is \frac{μ^{2}}{\code{shape}}. Thus, the constant multiplier here is \frac{1}{√{\code{shape}}}.

Similarly, for Inverse Gaussian fits, we have inverse.gaussian()$variance = mu^{3}, with known variance \frac{μ^{3}}{λ}; similarly, the constant multiplier here is \frac{1}{√{λ}}. λ is referred to as the shape parameter.

Value

The value differs based on the family (see details above for an explanation.)

For Gaussian fitted models, this is the residual standard deviation.

For Gamma and Inverse Gaussian fitted models, it represents the square root of the inverse of the shape parameter.

For Binomial and Poisson fitted models, the value is always reported as 1, since their variance is determined entirely by the mean and there is no separate scale parameter.

See Also

Package lme4 provides methods for mixed-effects models of class merMod and lists of linear models, lmList4.

Examples

methods(sigma)# from R 3.3.0 on, shows methods from pkgs 'stats' *and* 'lme4'

[Package lme4 version 2.0-1 Index]