summary {lme4}	R Documentation

Summary for a [ng]lmer Fit

Description

summary.merMod Computes and returns a list of summary statistics of a fitted mixed-effects model of class merMod, as produced by [ng]lmer fits. The summary includes estimates for the fixed and random effects. print.summary controls the output for the summary method.

Usage

## S3 method for class 'merMod'
summary(object,
       correlation = (p <= getOption("lme4.summary.cor.max")),
       use.hessian = NULL, ...)
## S3 method for class 'summary.merMod'
print(x, digits = max(3, getOption("digits") - 3),
      correlation = NULL, symbolic.cor = FALSE,
      signif.stars = getOption("show.signif.stars"),
      ranef.comp = c("Variance", "Std.Dev."),
      ranef.corr = any(ranef.comp == "Std.Dev."), show.resids = TRUE, ...)

Arguments

object
a fitted [ng]lmer model.

correlation
(logical) indicates whether the correlation matrix should be computed and stored along with the covariance. For print.summary.merMod, indicates whether the correlation matrix of the fixed-effects parameters should be printed. In the latter case, when NULL (the default), the correlation matrix is printed when it has been computed by summary(.), and when p <= 12, and the cutoff 12 may be modified by options(lme4.summary.cor.max = <n>)

use.hessian
(logical) indicates whether to use the finite-difference Hessian of the deviance function to compute standard errors of the fixed effects; see vcov.merMod for details

x
an R object of class summary.merMod.

digits
number of significant digits for printing.

symbolic.cor
should a symbolic encoding of the fixed-effects correlation matrix be printed? If so, the symnum function is used.

signif.stars
(logical) should significance stars be used?

ranef.comp
character vector of length one or two, indicating if random-effects parameters should be reported on the variance and/or standard deviation scale.

ranef.corr
(logical) print correlations (rather than covariances) of random effects?

show.resids
should the quantiles of the scaled residuals be printed?

...
potentially further arguments passed from other methods.

Value

An S3 object of class "summary.merMod", which is a list containing fixed and random effect estimates along with other standard model summary information.

See Also

summary for the original function in base R,

lmer, and glmer for model fitting,

sigma.merMod documents the result obtained in "Residual Std.Dev"; the name "Residual Std.Dev" is mostly applicable for lmer fits, but the meaning differs for glmer fits.

Examples

fm1 <- lmer(Reaction ~ Days + (Days | Subject), sleepstudy)
smry <- summary(fm1)
# Obtaining the variance-covariance matrix of the fixed effects
smry$vcov
# Obtaining the correlation matrix of fixed effects
smry$vcov@factors$correlation

[Package lme4 version 2.0-1 Index]