checkConv {lme4}	R Documentation

Extended Convergence Checking

Description

Primarily internal code for checking optimization convergence, see convergence for a more detailed discussion.

Usage

checkConv(derivs, coefs, ctrl, lbound, debug = FALSE)

Arguments

derivs
typically the "derivs" attribute of optimizeLmer(); with "gradients" and possibly "Hessian" component

coefs
current coefficient estimates

ctrl
list of lists, each with action character strings specifying what should happen when a check triggers, and tol numerical tolerances, as is the result of lmerControl()$checkConv.

lbound
vector of lower bounds for random-effects parameters only (length is taken to determine number of RE parameters)

debug
enable debugging output, useful if some checks are on "ignore", but would "trigger"

Value

A result list containing

code
The return code for the check

messages
A character vector of warnings and messages generated by the check

See Also

convergence

[Package lme4 version 1.1-30 Index]