devcomp {lme4}	R Documentation

Extract the deviance component list

Description

Return the deviance component list

Usage

  devcomp(x)

Arguments

x
a fitted model of class merMod

Details

A fitted model of class merMod has a devcomp slot as described in the value section.

Value

a list with components

dims
a named integer vector of various dimensions

cmp
a named numeric vector of components of the deviance

Note

This function is deprecated, use getME(., "devcomp")

[Package lme4 version 1.1-30 Index]