add information to data based on a fitted model

fortify adds information to data based on a fitted model; getData retrieves data as specified in the data argument

Usage

fortify.merMod(model, data = getData(model),
    ...)
# S3 method for merMod
getData(object)

Arguments

model

fitted model

object

fitted model

data

original data set, if needed

...

additional arguments

Details

fortify is defined in the ggplot2 package, q.v. for more details. fortify is not defined here, and fortify.merMod is defined as a function rather than an S3 method, to avoid (1) inducing a dependency on ggplot2 or (2) masking methods from ggplot2. This feature is both experimental and semi-deprecated, as the help page for fortify itself says: “Rather than using this function, I now recommend using the broom package, which implements a much wider range of methods. fortify may be deprecated in the future.” The broom.mixed package is recommended for mixed models in general.

getData is a bare-bones implementation; it relies on a data argument having been specified and the data being available in the environment of the formula. Unlike the functions in the nlme package, it does not do anything special with na.action or subset.

Examples

  fm1 <- lmer(Reaction~Days+(1|Subject),sleepstudy)
  names(fortify.merMod(fm1))
#> [1] "Reaction" "Days"     "Subject"  ".fitted"  ".resid"   ".scresid"