glmFamily {lme4}	R Documentation

Generator object for the glmFamily class

Description

The generator object for the glmFamily reference class. Such an object is primarily used through its new method.

Usage

  glmFamily(...)

Arguments

...
Named argument (see Note below)

Methods

new(family=family)

Create a new glmFamily object

Note

Arguments to the new method must be named arguments.

See Also

glmFamily

[Package lme4 version 1.1-31 Index]