Class "golden" and Generator for Golden Search Optimizer Class

"golden" is a reference class for a golden search scalar optimizer, for a parameter within an interval.

golden() is the generator for the "golden" class. The optimizer uses reverse communications.

Usage

golden(...)

Arguments

...

(partly optional) arguments passed to new() must be named arguments. lower and upper are the bounds for the scalar parameter; they must be finite.

Extends

All reference classes extend and inherit methods from "envRefClass".

Examples

showClass("golden")
#> Class "golden" [package "lme4"]
#> 
#> Slots:
#>                   
#> Name:       .xData
#> Class: environment
#> 
#> Extends: 
#> Class "envRefClass", directly
#> Class ".environment", by class "envRefClass", distance 2
#> Class "refClass", by class "envRefClass", distance 2
#> Class "environment", by class "envRefClass", distance 3, with explicit coerce
#> Class "refObject", by class "envRefClass", distance 3

golden(lower= -100, upper= 1e100)
#> Reference class object of class "golden"
#> Field "Ptr":
#> <pointer: 0x000001de5b8c2600>
#> Field "lowerbd":
#> [1] -100
#> Field "upperbd":
#> [1] 1e+100