hatvalues.merMod {lme4}	R Documentation

Diagonal elements of the hat matrix

Description

Returns the values on the diagonal of the hat matrix, which is the matrix that transforms the response vector (minus any offset) into the fitted values (minus any offset). Note that this method should only be used for linear mixed models. It is not clear if the hat matrix concept even makes sense for generalized linear mixed models.

Usage

  ## S3 method for class 'merMod'
 hatvalues(model, fullHatMatrix = FALSE, ...)

Arguments

model
An object of class merMod.

fullHatMatrix
Return full hat matrix (not just diagonal values)?

...
Not currently used

Value

The diagonal elements of the hat matrix.

Examples

m <- lmer(Reaction ~ Days + (Days | Subject), sleepstudy)
hatvalues(m)

[Package lme4 version 1.1-30 Index]