Class "lmList4" of 'lm' Objects on Common Model

Class "lmList4" is an S4 class with basically a list of objects of class lm with a common model (but different data); see lmList() which returns these.

Package nlme's lmList() returns objects of S3 class "lmList" and provides methods for them, on which our methods partly build.

Objects from the Class

Objects can be created by calls of the form new("lmList4", ...) or, more commonly, by a call to lmList().

Methods

A dozen methods are provided. Currently, S4 methods for show, coercion (as(.,.)) and others inherited via "list", and S3 methods for coef, confint, fitted, fixef, formula, logLik, pairs, plot, predict, print, qqnorm, ranef, residuals, sigma, summary, and update.

sigma(object)

returns the standard deviation \(\hat{\sigma}\) (of the errors in the linear models), assuming a common variance \(\sigma^2\) by pooling (even when pool = FALSE was used in the fit).

See also

Examples

if(getRversion() >= "3.2.0") {
  (mm <- methods(class = "lmList4"))
  ## The S3 ("not S4") ones :
  mm[!attr(mm,"info")[,"isS4"]]
}
#>  [1] "coef.lmList4"      "confint.lmList4"   "fitted.lmList4"   
#>  [4] "fixef.lmList4"     "formula.lmList4"   "logLik.lmList4"   
#>  [7] "pairs.lmList4"     "plot.lmList4"      "predict.lmList4"  
#> [10] "qqnorm.lmList4"    "ranef.lmList4"     "residuals.lmList4"
#> [13] "sigma.lmList4"     "summary.lmList4"   "update.lmList4"   
## For more examples:  example(lmList)  i.e., ?lmList