lmResp {lme4}	R Documentation

Generator objects for the response classes

Description

The generator objects for the lmResp, lmerResp, glmResp and nlsResp reference classes. Such objects are primarily used through their new methods.

Usage

  lmResp(...)

Arguments

...
List of arguments (see Note).

Methods

new(y=y):

Create a new lmResp or lmerResp object.

new(family=family, y=y):

Create a new glmResp object.

new(y=y, nlmod=nlmod, nlenv=nlenv, pnames=pnames, gam=gam):

Create a new nlsResp object.

Note

Arguments to the new methods must be named arguments.

y the numeric response vector

family a family object

nlmod the nonlinear model function

nlenv an environment holding data objects for evaluation of nlmod

pnames a character vector of parameter names

gam a numeric vector - the initial linear predictor

See Also

lmResp, lmerResp, glmResp, nlsResp

[Package lme4 version 1.1-30 Index]