mkMerMod {lme4}	R Documentation

Create a 'merMod' Object

Description

Create an object of (a subclass of) class merMod from the environment of the objective function and the value returned by the optimizer.

Usage

  mkMerMod(rho, opt, reTrms, fr, mc, lme4conv = NULL)

Arguments

rho
the environment of the objective function

opt
the optimization result returned by the optimizer (a list: see lmerControl for required elements)

reTrms
random effects structure from the calling function (see mkReTrms for required elements)

fr
model frame (see model.frame)

mc
matched call from the calling function

lme4conv
lme4-specific convergence information (results of checkConv)

Value

an object from a class that inherits from merMod.

[Package lme4 version 1.1-31 Index]