plot.lmList4 {lme4}	R Documentation

plots for lmList4 objects

Description

diagnostic and confidence-interval plots for lmList fits

Usage

## S3 method for class 'lmList4'
plot(x, form, abline, id, idLabels, grid, ...)
## S3 method for class 'lmList4.confint'
plot(x, y, order, ...)

Arguments

x
an object inheriting from class lmList, representing a list of lm objects with a common model.

y
ignored: for agreement with generic method

form
an optional formula specifying the desired type of plot. Any variable present in the original data frame used to obtain x can be referenced. In addition, x itself can be referenced in the formula using the symbol ".". Conditional expressions on the right of a | operator can be used to define separate panels in a Trellis display. Default is resid(., type = "pool") ~ fitted(.) , corresponding to a plot of the standardized residuals (using a pooled estimate for the residual standard error) versus fitted values.

abline
an optional numeric value, or numeric vector of length two. If given as a single value, a horizontal line will be added to the plot at that coordinate; else, if given as a vector, its values are used as the intercept and slope for a line added to the plot. If missing, no lines are added to the plot.

id
an optional numeric value, or one-sided formula. If given as a value, it is used as a significance level for a two-sided outlier test for the standardized residuals. Observations with absolute standardized residuals greater than the 1 - value/2 quantile of the standard normal distribution are identified in the plot using idLabels. If given as a one-sided formula, its right hand side must evaluate to a logical, integer, or character vector which is used to identify observations in the plot. If missing, no observations are identified.

idLabels
an optional vector, or one-sided formula. If given as a vector, it is converted to character and used to label the observations identified according to id. If given as a one-sided formula, its right hand side must evaluate to a vector which is converted to character and used to label the identified observations. Default is getGroups(x).

grid
an optional logical value indicating whether a grid should be added to plot. Default depends on the type of Trellis plot used: if xyplot defaults to TRUE, else defaults to FALSE.

order
which coefficient to order the results by

...
optional arguments passed to the Trellis plot function.

Details

The plot method for lmList4 objects is copied from that for lmList objects; see plot.lmList for details.

Author(s)

Original versions in nlme package by Jose Pinheiro and Douglas Bates

Examples

fm.plm  <- lmList(Reaction ~ Days | Subject, sleepstudy)
## diagnostic plot: standardized residuals vs. fitted
plot(fm.plm,id=0.05)
ci <- confint(fm.plm)
## plot CIs, ordered by slope (coefficient 2)
plot(ci,order=2,ylab="Subject")

[Package lme4 version 1.1-15 Index]