plots.thpr {lme4}	R Documentation

Mixed-Effects Profile Plots (Regular / Density / Pairs)

Description

Xyplot, Densityplot, and Pairs plot methods for a mixed-effects model profile.

xyplot() draws “zeta diagrams”, also visualizing confidence intervals and their asymmetry.

densityplot() draws the profile densities.

splom() draws profile pairs plots. Contours are for the marginal two-dimensional regions (i.e. using df = 2).

Usage

## S3 method for class 'thpr'
xyplot(x, data = NULL,
       levels = sqrt(qchisq(pmax.int(0, pmin.int(1, conf)), df = 1)),
       conf = c(50, 80, 90, 95, 99)/100,
       absVal = FALSE, scales=NULL,
       which = 1:nptot, ...)

## S3 method for class 'thpr'
densityplot(x, data, npts = 201, upper = 0.999, ...)

## S3 method for class 'thpr'
splom(x, data,
      levels = sqrt(qchisq(pmax.int(0, pmin.int(1, conf)), 2)),
      conf = c(50, 80, 90, 95, 99)/100,  which = 1:nptot,
      draw.lower = TRUE, draw.upper = TRUE, ...)

Arguments

x
a mixed-effects profile, i.e., of class "thpr", typically resulting from profile(fm) where fm is a fitted model from lmer (or its generalizations).

data
unused - only for compatibility with generic.

npts
the number of points to use for the densityplot().

upper
a number in (0,1) to specify upper (and lower) boundaries as +/- qnorm(upper).

levels
the contour levels to be shown; usually derived from conf.

conf
numeric vector of confidence levels to be shown as contours.

absVal
logical indicating if abs(.)olute values should be plotted, often preferred for confidence interval visualization.

scales
plotting options to be passed to xyplot

which
integer or character vector indicating which parameters to profile: default is all parameters (see profile-methods for details).

draw.lower
(logical) draw lower-triangle (zeta scale) panels?

draw.upper
(logical) draw upper-triangle (standard dev/cor scale) panels?

...
further arguments passed to xyplot, densityplot, or splom from package lattice, respectively.

Value

xyplot:

a density plot, a "trellis" object (lattice package) which when print()ed produces plots on the current graphic device.

densityplot:

a density plot, a "trellis" object, see above.

splom:

a pairs plot, aka scatterplot matrix, a "trellis" object, see above.

See Also

profile, notably for an example.

Examples

## see   example("profile.merMod")

[Package lme4 version 1.1-31 Index]