rePos {lme4}	R Documentation

Generator object for the rePos (random-effects positions) class

Description

The generator object for the rePos class used to determine the positions and orders of random effects associated with particular random-effects terms in the model.

Usage

  rePos(...)

Arguments

...
Argument list (see Note).

Methods

new(mer=mer)

Create a new rePos object.

Note

Arguments to the new methods must be named arguments. mer, an object of class "merMod", is the only required/expected argument.

See Also

rePos

[Package lme4 version 1.1-31 Index]