refitML {lme4}	R Documentation

Refit a Model by Maximum Likelihood Criterion

Description

Refit a (merMod) model using the maximum likelihood criterion.

Usage

refitML(x, ...)
## S3 method for class 'merMod'
 refitML(x, optimizer = "bobyqa", ...)

Arguments

x
a fitted model, usually of class "lmerMod", to be refit according to the maximum likelihood criterion.

...
optional additional parameters. None are used at present.

optimizer
a string indicating the optimizer to be used.

Details

This function is primarily used to get a maximum likelihood fit of a linear mixed-effects model for an anova comparison.

Value

an object like x but fit by maximum likelihood

See Also

refit and update.merMod for more extensive refitting.

[Package lme4 version 1.1-31 Index]