sigma {lme4}	R Documentation

Extract Residual Standard Deviation 'Sigma'

Description

Extract the estimated standard deviation of the errors, the “residual standard deviation” (also misnamed the “residual standard error”), from a fitted model.

Usage

## S3 method for class 'merMod'
sigma(object, ...)

Arguments

object
a fitted model.

...
additional, optional arguments, passed from or to methods. (None currently in our two methods.)

Details

Package lme4 provides methods for mixed-effects models of class merMod and lists of linear models, lmList4.

Value

Typically a number, the estimated standard deviation of the errors (“residual standard deviation”) for Gaussian models, and - less interpretably - the square root of the residual deviance per degree of freedom in more general models.

Examples

methods(sigma)# from R 3.3.0 on, shows methods from pkgs 'stats' *and* 'lme4'

[Package lme4 version 1.1-30 Index]