Extract Residual Standard Deviation 'Sigma'

Extract the estimated standard deviation of the errors, the “residual standard deviation” (also misnamed the “residual standard error”), from a fitted model.

Usage

# S3 method for merMod
sigma(object, ...)

Arguments

object

a fitted model.

...

additional, optional arguments, passed from or to methods. (None currently in our two methods.)

Value

Typically a number, the estimated standard deviation of the errors (“residual standard deviation”) for Gaussian models, and - less interpretably - the square root of the residual deviance per degree of freedom in more general models.

Details

Package lme4 provides methods for mixed-effects models of class merMod and lists of linear models, lmList4.

Examples

methods(sigma)# from R 3.3.0 on, shows methods from pkgs 'stats' *and* 'lme4'
#> [1] sigma.default* sigma.gls*     sigma.lmList*  sigma.lmList4* sigma.lme*    
#> [6] sigma.merMod*  sigma.mlm*    
#> see '?methods' for accessing help and source code